requirements.txt
allure-pytest
playwright
python-dotenv
requests
